+++ xdis/unmarshal.py
@@ -236,9 +236,9 @@ def load_code_type(fp, magic_int, bytes_for_s=False, code_objects={}):
                     co_cellvars,
                 )
             else:
-                if PYTHON_MAGIC_INT in (3412, 3413):
+                if PYTHON_MAGIC_INT in (3412, 3413, 3422):
                     if co_posonlyargcount is not None:
-                        # Python3.8 to Python3.8: Ok to use native Python3.8's code type
+                        # Python3.8 to Python3.9: Ok to use native Python3.8's code type
                         code = Code(
                             co_argcount,
                             co_posonlyargcount,
@@ -278,6 +278,8 @@ def load_code_type(fp, magic_int, bytes_for_s=False, code_objects={}):
                             co_cellvars,
                         )
                 elif co_posonlyargcount is not None:
+                    if not isinstance(co_lnotab, bytes):
+                        co_lnotab = bytes(co_lnotab, encoding="utf-8")
                     code = Code38(
                         co_argcount,
                         co_posonlyargcount,
@@ -292,7 +294,7 @@ def load_code_type(fp, magic_int, bytes_for_s=False, code_objects={}):
                         co_filename,
                         co_name,
                         co_firstlineno,
-                        bytes(co_lnotab, encoding="utf-8"),
+                        co_lnotab,
                         co_freevars,
                         co_cellvars,
                     )